[PATCH] 64-bit sector_t - filesystems
From Peter Chubb

Filesystem migration to possibly 64-bit sector_t:
 - bmap() now takes and returns a sector_t to allow filesystems
   (e.g., JFS, XFS) that are 64-bit clean to deal with large files
 - buffer handling now 64-bit clean

Enable 64-bit sector_t on IA32 and PPC.

kiobufs takes sector_t array, not array of long.
Fix blkmtd.c to deal in such an array.

Miscellaneous fixes for 64-bit sector_t.
 	 - missed printk formats
	 - ide_floppy_do_request had incorrect signature
	 - in blkmtd.c there was a pointer used to
	   manipulate an array to be used by kiobuf --
 	   it was unsigned long, needed to be sector_t

+++ fs/ext3/ialloc.c
@@ -479,7 +479,7 @@ struct inode *ext3_orphan_get (struct super_block * sb, ino_t ino)
 			!(inode = iget(sb, ino)) || is_bad_inode(inode) ||
 			NEXT_ORPHAN(inode) > max_ino) {
 		ext3_warning(sb, __FUNCTION__,
-			     "bad orphan inode %lu!  e2fsck was run?\n", ino);
+			     "bad orphan inode %lu!  e2fsck was run?\n", (unsigned long)ino);
 		printk(KERN_NOTICE "ext3_test_bit(bit=%d, block=%llu) = %d\n",
 		       bit, 
 			(unsigned long long)bitmap_bh->b_blocknr,
